moves layout & query highlight stores
