+++ src/lib/application-state-stores/query-highlight-store.ts
