+++ src/routes/_surfaces/inspector/index.svelte
@@ -6,7 +6,7 @@ import type { ApplicationStore } from "$lib/application-state-stores/application
 
 import { drag } from "$lib/drag";
 import { EntityType } from "$common/data-modeler-state-service/entity-state-service/EntityStateService";
-import { inspectorVisibilityTween, inspectorVisible, layout } from "$lib/layout-store";
+import { inspectorVisibilityTween, inspectorVisible, layout } from "$lib/application-state-stores/layout-store";
 import Portal from "$lib/components/Portal.svelte";
 
 const store = getContext('rill:app:store') as ApplicationStore;
